wip towards a clean tracker package
