+++ src/levanter/tracker/__init__.py
@@ -1,0 +1,9 @@
+from levanter.tracker.helpers import log_optimizer_hyperparams
+from levanter.tracker.tracker import (
+    CompositeTracker,
+    Tracker,
+    current_tracker,
+    jit_log_metrics,
+    log_metrics,
+    log_summary,
+)
