+++ src/levanter/trainer.py
@@ -31,6 +31,8 @@ from haliax.partitioning import ResourceAxis, ResourceMapping, named_jit
 
 import levanter.logging
 import levanter.tracker
+import levanter.tracker.tracker
+import levanter.tracker.wandb
 from levanter import logging
 from levanter.checkpoint import CheckpointerConfig
 from levanter.config import JsonAtom
@@ -38,7 +40,7 @@ from levanter.data import Dataset, ReplicatedBatchLoader, ShardableDataset, Shar
 from levanter.distributed import DistributedConfig, RayConfig
 from levanter.grad_accum import accumulate_gradients_sharded
 from levanter.logging import capture_time
-from levanter.tracker import WandbConfig
+from levanter.tracker.wandb import WandbConfig
 from levanter.types import FilterSpec
 from levanter.utils import cloud_utils
 from levanter.utils.jax_utils import is_inexact_arrayish
@@ -117,7 +119,7 @@ class Trainer:
     config: "TrainerConfig"
     optimizer: GradientTransformation
     hooks: TrainerHooks
-    _tracker: levanter.tracker.Tracker
+    _tracker: levanter.tracker.tracker.Tracker
     is_trainable_param: Optional[PyTree[FilterSpec]]
     _raw_loss_function: Callable
     _cmanagers: List[typing.ContextManager] = []
@@ -146,9 +148,8 @@ class Trainer:
         self._raw_loss_function = loss_fn
         self.optimizer = optimizer
         self.is_trainable_param = is_trainable
-        self._tracker = levanter.tracker.WandbTracker(self.config.wandb)
         # TODO: hacky hack
-        self._tracker._run = wandb.run
+        self._tracker = levanter.tracker.wandb.WandbTracker(wandb.run)
         self._cmanagers = []
 
     @cached_property
@@ -216,7 +217,7 @@ class Trainer:
             raise RuntimeError("Trainer is already entered")
 
         self._cmanagers = [
-            levanter.tracker.current_tracker(self._tracker),
+            levanter.current_tracker(self._tracker),
             self.device_mesh,
             hax.axis_mapping(self.parameter_axis_mapping),
         ]
@@ -292,7 +293,7 @@ class Trainer:
         """
         Performs a single training step.
         """
-        with capture_time() as step_time, levanter.tracker.current_tracker(self._tracker):
+        with capture_time() as step_time, levanter.current_tracker(self._tracker):
             key, new_key = jax.random.split(state.training_key)
             loss, new_model, new_optstate = self._train_step_fn(
                 state.model, state.opt_state, *batch, **batch_kwargs, key=key
@@ -309,13 +310,12 @@ class Trainer:
         Generator that yields training steps and runs hooks.
         """
         iter_data = iter(train_loader)
-        with levanter.tracker.current_tracker(self._tracker):
+        with levanter.current_tracker(self._tracker):
             while state.step < self.config.num_train_steps:
                 with capture_time() as loading_time:
                     example = next(iter_data)
 
-                # TODO: refactor logging
-                levanter.tracker.log_metrics({"throughput/loading_time": loading_time()}, step=state.step)
+                levanter.log_metrics({"throughput/loading_time": loading_time()}, step=state.step)
 
                 info = self.train_step(state, example)
                 state = info.state
@@ -324,7 +324,7 @@ class Trainer:
                     with capture_time() as hook_time:
                         self.run_hooks(info)
 
-                    levanter.tracker.log_metrics({"throughput/hook_time": hook_time()}, step=state.step)
+                    levanter.log_metrics({"throughput/hook_time": hook_time()}, step=state.step)
 
                 yield info
