+++ tests/test_viz_lm.py
@@ -12,7 +12,7 @@ import tiny_test_corpus
 from levanter.checkpoint import save_checkpoint
 from levanter.distributed import RayConfig
 from levanter.models.gpt2 import Gpt2Config, Gpt2LMHeadModel
-from levanter.tracker import WandbConfig
+from levanter.tracker.wandb import WandbConfig
 from levanter.utils.py_utils import logical_cpu_core_count
